+++ extensions/obsidian-smart-capture/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Obsidian Smart Capture Changelog
 
+## [New Features] -- 2025-25-10
+- Opens Obsidian in the background (does not capture screen). Toggled in settings
+- Changed the format of notes being saved
+- Allows user the option to open notes in new tabs
+- Allows user to keep appending information to the same note if the same filename is used.
+
 ## [New features] - 2025-03-10
 
 - Add support for generating summary using AI